moves template logic to form component
